% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdown.R
\docType{package}
\name{beardown}
\alias{beardown}
\alias{beardown-package}
\title{beardown: A package for creating PhD theses for Baylor University
using R Markdown, based on the thesisdown package}
\description{
beardown: A package for creating PhD theses for Baylor University
using R Markdown, based on the thesisdown package
}
\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

